First implementation of ERP - Waits for the first mainmenu load for SteamAPI to initialize - Mod can now detect if Online or LAN mode - ERP is sending to Steam currently
Mod doesn't work since the game developer has to setup localization for the Rich Presenece keys. The only thing that works is the `steam_player_group`.

"steam_display" - Names a rich presence localization token that will be displayed in the viewing user's selected language in the Steam client UI. See Rich Presence Localization for more info, including a link to a page for testing this rich presence data. If steam_display is not set to a valid localization tag, then rich presence will not be displayed in the Steam client.
Source: https://partner.steamgames.com/doc/api/ISteamFriends#SetRichPresence